Minor refactor
Split out index to cma.ts, move message tests to their own file, move
typeGuards to their own file
